requirements.txt
amqp==5.2.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
black==24.4.2
blessed==1.20.0
bpython==0.24
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
curtsies==0.4.2
cwcwidth==0.1.9
decorator==5.1.1
Django==5.0.6
django-celery-beat==2.6.0
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
executing==2.0.1
Faker==25.2.0
flake8==7.0.0
flake8-quotes==3.4.0
greenlet==3.0.3
idna==3.7
inflection==0.5.1
ipython==8.24.0
jedi==0.19.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.2.2
prompt-toolkit==3.0.43
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.8.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
redis==5.0.4
referencing==0.35.1
requests==2.32.2
rpds-py==0.18.1
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
